feat(api_server): change volume_id from u32 to u16 throughout codebase
- Update all RPC client function signatures to use u16 for volume_id
- Move DataVgInfo and DataVolume from protobuf to native Rust structs
  since protobuf doesn't support u16 types
- Ensure consistent u16 usage across data and metadata volume proxies
- Print data blob guid as {blob_id}:d{volume_id} and metadata blob guid
  as {blob_id}:m{volume_id}

+++ crates/api_server/src/blob_storage.rs
@@ -341,7 +341,7 @@ impl BlobStorageImpl {
         &self,
         tracking_root_blob_name: Option<&str>,
         blob_id: Uuid,
-        volume_id: u32,
+        volume_id: u16,
         block_number: u32,
         body: Bytes,
     ) -> Result<DataBlobGuid, BlobStorageError> {
